ai-surrogate-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.3.0
google-generativeai==0.3.2
openai-whisper==20231117
gtts==2.5.1
langchain==0.1.0
langgraph==0.0.20
asyncio==3.4.3
aiofiles==23.2.1
websockets==12.0
requests==2.31.0
numpy==1.24.3
torch==2.1.0
librosa==0.10.1
soundfile==1.0.7
mutagen==1.47.0